refactor: change directory of budget module
